// service/types.gen.ts
export type SpaceResponseSchema = {
    id: string;
    type: string;
    setting: {
        [key: string]: string;
    };
    count: number;
    state: string;
    /**
     * 价格因子
     */
    price_factor?: number;
    updatedAt: string;
    shopId: string;
    shop?: {
        /**
         * 商标
         */
        trademark?: string;
        /**
         * 门店号
         */
        shop_no?: string;
        branch?: string;
        type_tag?: string;
    };
    /**
     * 分类标签
     */
    tag?: string;
    /**
     * 位置，1-主客区/大堂 2-商家入口 3-入口通道 4-独立房间/包间 5-通往洗手间过道 6-洗手间 7-商家外摆区/店外公共区 8-店外墙面(非临街) 9-店外墙面(临街)
     */
    site?: string;
    /**
     * 稳定性，1-固定 2-半固定 3-移动 4-临时
     */
    stability?: string;
    /**
     * 相册
     */
    photo?: Array<string>;
    /**
     * 投放推介
     */
    description?: string;
    /**
     * 设计注意事项
     */
    design_attention?: string;
    /**
     * 施工注意事项
     */
    construction_attention?: string;
};